+++ backends/rapidpro/backend_test.go
@@ -470,56 +470,94 @@ func (ts *BackendTestSuite) TestMsgStatus() {
 	now := time.Now().In(time.UTC)
 	time.Sleep(2 * time.Millisecond)
 
-	// update by id with external id
+	// put test message back into queued state
+	ts.b.db.MustExec(`UPDATE msgs_msg SET status = 'Q', sent_on = NULL WHERE id = $1`, 10001)
+
+	// update to WIRED using id and provide new external ID
 	status := ts.b.NewMsgStatusForID(channel, courier.NewMsgID(10001), courier.MsgWired)
 	status.SetExternalID("ext0")
 	err := ts.b.WriteMsgStatus(ctx, status)
 	ts.NoError(err)
 	time.Sleep(time.Second)
 
-	m, err := readMsgFromDB(ts.b, courier.NewMsgID(10001))
-	ts.NoError(err)
-	ts.Equal(m.Status_, courier.MsgWired)
-	ts.Equal(m.ExternalID_, null.String("ext0"))
+	m := readMsgFromDB(ts.b, courier.NewMsgID(10001))
+	ts.Equal(courier.MsgWired, m.Status_)
+	ts.Equal(null.String("ext0"), m.ExternalID_)
 	ts.True(m.ModifiedOn_.After(now))
 	ts.True(m.SentOn_.After(now))
 
-	// update by id, no external id, shouldn't overwrite it
+	sentOn := *m.SentOn_
+
+	// update to SENT using id
 	status = ts.b.NewMsgStatusForID(channel, courier.NewMsgID(10001), courier.MsgSent)
 	err = ts.b.WriteMsgStatus(ctx, status)
 	ts.NoError(err)
 	time.Sleep(time.Second)
 
-	// no change for incoming messages
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10002))
-	ts.NoError(err)
-	ts.Equal(m.Status_, courier.MsgPending)
-	ts.Equal(m.ExternalID_, null.String("ext2"))
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10001))
+	ts.Equal(courier.MsgSent, m.Status_)
+	ts.Equal(null.String("ext0"), m.ExternalID_) // no change
+	ts.True(m.ModifiedOn_.After(now))
+	ts.True(m.SentOn_.Equal(sentOn)) // no change
 
+	// update to DELIVERED using id
+	status = ts.b.NewMsgStatusForID(channel, courier.NewMsgID(10001), courier.MsgDelivered)
+	err = ts.b.WriteMsgStatus(ctx, status)
+	ts.NoError(err)
+	time.Sleep(time.Second)
+
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10001))
+	ts.Equal(m.Status_, courier.MsgDelivered)
+	ts.True(m.ModifiedOn_.After(now))
+	ts.True(m.SentOn_.Equal(sentOn)) // no change
+
+	// no change for incoming messages
 	status = ts.b.NewMsgStatusForID(channel, courier.NewMsgID(10002), courier.MsgSent)
 	err = ts.b.WriteMsgStatus(ctx, status)
 	ts.NoError(err)
 	time.Sleep(time.Second)
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10002))
-	ts.NoError(err)
-	ts.Equal(m.Status_, courier.MsgPending)
+
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10002))
+	ts.Equal(courier.MsgPending, m.Status_)
 	ts.Equal(m.ExternalID_, null.String("ext2"))
 
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10001))
-	ts.NoError(err)
-	ts.Equal(m.Status_, courier.MsgSent)
-	ts.Equal(m.ExternalID_, null.String("ext0"))
-	ts.True(m.ModifiedOn_.After(now))
-
-	// update by external id
+	// update to FAILED using external id
 	status = ts.b.NewMsgStatusForExternalID(channel, "ext1", courier.MsgFailed)
 	err = ts.b.WriteMsgStatus(ctx, status)
 	ts.NoError(err)
 	time.Sleep(time.Second)
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10000))
-	ts.NoError(err)
-	ts.Equal(m.Status_, courier.MsgFailed)
+
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10000))
+	ts.Equal(courier.MsgFailed, m.Status_)
 	ts.True(m.ModifiedOn_.After(now))
+	ts.Nil(m.SentOn_)
+
+	now = time.Now().In(time.UTC)
+	time.Sleep(2 * time.Millisecond)
+
+	// update to WIRED using external id
+	status = ts.b.NewMsgStatusForExternalID(channel, "ext1", courier.MsgWired)
+	err = ts.b.WriteMsgStatus(ctx, status)
+	ts.NoError(err)
+	time.Sleep(time.Second)
+
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10000))
+	ts.Equal(courier.MsgWired, m.Status_)
+	ts.True(m.ModifiedOn_.After(now))
+	ts.True(m.SentOn_.After(now))
+
+	sentOn = *m.SentOn_
+
+	// update to SENT using external id
+	status = ts.b.NewMsgStatusForExternalID(channel, "ext1", courier.MsgSent)
+	err = ts.b.WriteMsgStatus(ctx, status)
+	ts.NoError(err)
+	time.Sleep(time.Second)
+
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10000))
+	ts.Equal(courier.MsgSent, m.Status_)
+	ts.True(m.ModifiedOn_.After(now))
+	ts.True(m.SentOn_.Equal(sentOn)) // no change
 
 	// no such external id for outgoing message
 	status = ts.b.NewMsgStatusForExternalID(channel, "ext2", courier.MsgSent)
@@ -534,6 +572,7 @@ func (ts *BackendTestSuite) TestMsgStatus() {
 	// reset our status to sent
 	status = ts.b.NewMsgStatusForExternalID(channel, "ext1", courier.MsgSent)
 	err = ts.b.WriteMsgStatus(ctx, status)
+	ts.NoError(err)
 	time.Sleep(time.Second)
 
 	// error our msg
@@ -544,8 +583,7 @@ func (ts *BackendTestSuite) TestMsgStatus() {
 	ts.NoError(err)
 	time.Sleep(time.Second)
 
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10000))
-	ts.NoError(err)
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10000))
 	ts.Equal(m.Status_, courier.MsgErrored)
 	ts.Equal(m.ErrorCount_, 1)
 	ts.True(m.ModifiedOn_.After(now))
@@ -557,8 +595,7 @@ func (ts *BackendTestSuite) TestMsgStatus() {
 	ts.NoError(err)
 	time.Sleep(time.Second)
 
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10000))
-	ts.NoError(err)
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10000))
 	ts.Equal(m.Status_, courier.MsgErrored)
 	ts.Equal(m.ErrorCount_, 2)
 
@@ -568,8 +605,7 @@ func (ts *BackendTestSuite) TestMsgStatus() {
 	time.Sleep(time.Second)
 
 	ts.NoError(err)
-	m, err = readMsgFromDB(ts.b, courier.NewMsgID(10000))
-	ts.NoError(err)
+	m = readMsgFromDB(ts.b, courier.NewMsgID(10000))
 	ts.Equal(m.Status_, courier.MsgFailed)
 	ts.Equal(m.ErrorCount_, 3)
 
@@ -674,7 +710,7 @@ func (ts *BackendTestSuite) TestDupes() {
 	uuid2 := msg.UUID().String()
 
 	// an outgoing message should clear things
-	dbMsg, err := readMsgFromDB(ts.b, courier.NewMsgID(10000))
+	dbMsg := readMsgFromDB(ts.b, courier.NewMsgID(10000))
 	dbMsg.URN_ = urn
 	dbMsg.channel = knChannel
 	dbMsg.ChannelUUID_ = knChannel.UUID()
@@ -718,8 +754,7 @@ func (ts *BackendTestSuite) TestExternalIDDupes() {
 
 func (ts *BackendTestSuite) TestLoop() {
 	ctx := context.Background()
-	dbMsg, err := readMsgFromDB(ts.b, courier.NewMsgID(10000))
-	ts.NoError(err)
+	dbMsg := readMsgFromDB(ts.b, courier.NewMsgID(10000))
 
 	dbMsg.ResponseToID_ = courier.MsgID(5)
 
@@ -755,9 +790,8 @@ func (ts *BackendTestSuite) TestStatus() {
 	r := ts.b.redisPool.Get()
 	defer r.Close()
 
-	dbMsg, err := readMsgFromDB(ts.b, courier.NewMsgID(10000))
+	dbMsg := readMsgFromDB(ts.b, courier.NewMsgID(10000))
 	dbMsg.ChannelUUID_, _ = courier.NewChannelUUID("dbc126ed-66bc-4e28-b67b-81dc3327c95d")
-	ts.NoError(err)
 	ts.NotNil(dbMsg)
 
 	// serialize our message
@@ -777,9 +811,8 @@ func (ts *BackendTestSuite) TestOutgoingQueue() {
 	r := ts.b.redisPool.Get()
 	defer r.Close()
 
-	dbMsg, err := readMsgFromDB(ts.b, courier.NewMsgID(10000))
+	dbMsg := readMsgFromDB(ts.b, courier.NewMsgID(10000))
 	dbMsg.ChannelUUID_, _ = courier.NewChannelUUID("dbc126ed-66bc-4e28-b67b-81dc3327c95d")
-	ts.NoError(err)
 	ts.NotNil(dbMsg)
 
 	// serialize our message
@@ -814,8 +847,7 @@ func (ts *BackendTestSuite) TestOutgoingQueue() {
 	ts.Nil(err)
 
 	// checking another message should show unsent
-	msg3, err := readMsgFromDB(ts.b, courier.NewMsgID(10001))
-	ts.NoError(err)
+	msg3 := readMsgFromDB(ts.b, courier.NewMsgID(10001))
 	sent, err = ts.b.WasMsgSent(ctx, msg3.ID())
 	ts.NoError(err)
 	ts.False(sent)
@@ -975,8 +1007,8 @@ func (ts *BackendTestSuite) TestWriteAttachment() {
 	}
 
 	// load it back from the id
-	m, err := readMsgFromDB(ts.b, msg.ID())
-	ts.NoError(err)
+	m := readMsgFromDB(ts.b, msg.ID())
+
 	if ts.Equal(1, len(m.Attachments())) {
 		ts.True(strings.HasPrefix(m.Attachments()[0], "image/png:"))
 		ts.True(strings.HasSuffix(m.Attachments()[0], ".png"))
@@ -1008,8 +1040,7 @@ func (ts *BackendTestSuite) TestWriteMsg() {
 	ts.NotZero(msg.ID)
 
 	// load it back from the id
-	m, err := readMsgFromDB(ts.b, msg.ID())
-	ts.NoError(err)
+	m := readMsgFromDB(ts.b, msg.ID())
 
 	tx, err := ts.b.db.Beginx()
 	ts.NoError(err)
@@ -1124,8 +1155,7 @@ func (ts *BackendTestSuite) TestPreferredChannelCheckRole() {
 	time.Sleep(1 * time.Second)
 
 	// load it back from the id
-	m, err := readMsgFromDB(ts.b, msg.ID())
-	ts.NoError(err)
+	m := readMsgFromDB(ts.b, msg.ID())
 
 	tx, err := ts.b.db.Beginx()
 	ts.NoError(err)
@@ -1265,4 +1295,26 @@ func TestBackendSuite(t *testing.T) {
 
 type ServerTestSuite struct {
 	suite.Suite
+}
+
+// for testing only, returned DBMsg object is not fully populated
+func readMsgFromDB(b *backend, id courier.MsgID) *DBMsg {
+	m := &DBMsg{
+		ID_: id,
+	}
+	err := b.db.Get(m, selectMsgSQL, id)
+	if err != nil {
+		panic(err)
+	}
+
+	ch := &DBChannel{
+		ID_: m.ChannelID_,
+	}
+	err = b.db.Get(ch, selectChannelSQL, m.ChannelID_)
+	if err != nil {
+		panic(err)
+	}
+
+	m.channel = ch
+	return m
 }